requirements.txt
rel==0.4.9
websocket-client==1.6.3
websockets==11.0.3
